fix: resolve scoped access PR gates
