+++ apps/sva-studio-react/src/routes/admin/-iam-dsr-detail-page.test.tsx
@@ -18,6 +18,10 @@ vi.mock('../../providers/auth-provider', () => ({
   useAuth: () => useAuthMock(),
 }));
 
+vi.mock('../../providers/effective-access-provider', () => ({
+  useEffectiveAuth: () => useAuthMock(),
+}));
+
 vi.mock('../../lib/iam-api', () => ({
   getAdminDsrCase: (...args: unknown[]) => getAdminDsrCaseMock(...args),
 }));
