+++ apps/sva-studio-react/src/routes/admin/modules/-modules-page.test.tsx
@@ -15,6 +15,10 @@ vi.mock('../../../providers/auth-provider', () => ({
   useAuth: () => useAuthMock(),
 }));
 
+vi.mock('../../../providers/effective-access-provider', () => ({
+  useEffectiveAuth: () => useAuthMock(),
+}));
+
 vi.mock('../../../components/ConfirmDialog', () => ({
   ConfirmDialog: ({
     open,
