+++ packages/routing/src/app-route-bindings.ts
@@ -1,0 +1,69 @@
+import type { RouteComponent } from '@tanstack/react-router';
+
+export type AppRouteBindings = {
+  readonly home: RouteComponent;
+  readonly account: RouteComponent;
+  readonly accountPrivacy: RouteComponent;
+  readonly accountPrivacyDetail: RouteComponent;
+  readonly accountRules: RouteComponent;
+  readonly content: RouteComponent;
+  readonly contentCreate: RouteComponent;
+  readonly contentDetail: RouteComponent;
+  readonly mediaUsage: RouteComponent;
+  readonly newsList: RouteComponent;
+  readonly newsDetail: RouteComponent;
+  readonly newsEditor: RouteComponent;
+  readonly eventsList: RouteComponent;
+  readonly eventsDetail: RouteComponent;
+  readonly eventsEditor: RouteComponent;
+  readonly genericItemsList: RouteComponent;
+  readonly genericItemsDetail: RouteComponent;
+  readonly faqList: RouteComponent;
+  readonly faqDetail: RouteComponent;
+  readonly faqEditor: RouteComponent;
+  readonly cockpitCardsList: RouteComponent;
+  readonly cockpitCardsDetail: RouteComponent;
+  readonly cockpitCardsEditor: RouteComponent;
+  readonly projectsList: RouteComponent;
+  readonly projectsDetail: RouteComponent;
+  readonly projectsEditor: RouteComponent;
+  readonly genericItemsEditor: RouteComponent;
+  readonly poiList: RouteComponent;
+  readonly poiDetail: RouteComponent;
+  readonly poiEditor: RouteComponent;
+  readonly media: RouteComponent;
+  readonly adminMedia: RouteComponent;
+  readonly categories: RouteComponent;
+  readonly app: RouteComponent;
+  readonly interfaces: RouteComponent;
+  readonly help: RouteComponent;
+  readonly support: RouteComponent;
+  readonly license: RouteComponent;
+  readonly adminUsers: RouteComponent;
+  readonly adminUserCreate: RouteComponent;
+  readonly adminUserDetail: RouteComponent;
+  readonly adminOrganizations: RouteComponent;
+  readonly adminOrganizationCreate: RouteComponent;
+  readonly adminOrganizationDetail: RouteComponent;
+  readonly adminInstances: RouteComponent;
+  readonly adminInstanceCreate: RouteComponent;
+  readonly adminInstanceSetup: RouteComponent;
+  readonly adminInstanceDetail: RouteComponent;
+  readonly adminRoles: RouteComponent;
+  readonly adminRoleCreate: RouteComponent;
+  readonly adminRoleDetail: RouteComponent;
+  readonly adminGroups: RouteComponent;
+  readonly adminGroupCreate: RouteComponent;
+  readonly adminGroupDetail: RouteComponent;
+  readonly adminLegalTexts: RouteComponent;
+  readonly adminLegalTextCreate: RouteComponent;
+  readonly adminLegalTextDetail: RouteComponent;
+  readonly adminIam: RouteComponent;
+  readonly adminIamGovernanceDetail: RouteComponent;
+  readonly adminIamDsrDetail: RouteComponent;
+  readonly modules: RouteComponent;
+  readonly monitoring: RouteComponent;
+  readonly monitoringJobs: RouteComponent;
+  readonly monitoringJobDetail: RouteComponent;
+  readonly adminApiPhase1Test: RouteComponent;
+};
